+++ oap-server/server-storage-plugin/storage-iotdb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/iotdb/profile/IoTDBProfileTaskQueryDAO.java
@@ -34,6 +34,7 @@ import org.apache.skywalking.oap.server.core.storage.profiling.trace.IProfileTas
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.oap.server.storage.plugin.iotdb.IoTDBClient;
 import org.apache.skywalking.oap.server.storage.plugin.iotdb.IoTDBIndexes;
+import org.apache.skywalking.oap.server.storage.plugin.iotdb.utils.IoTDBUtils;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -46,22 +47,28 @@ public class IoTDBProfileTaskQueryDAO implements IProfileTaskQueryDAO {
                                          Long endTimeBucket, Integer limit) throws IOException {
         StringBuilder query = new StringBuilder();
         query.append("select * from ");
-        query = client.addModelPath(query, ProfileTaskRecord.INDEX_NAME);
+        IoTDBUtils.addModelPath(client.getStorageGroup(), query, ProfileTaskRecord.INDEX_NAME);
         Map<String, String> indexAndValueMap = new HashMap<>();
         if (StringUtil.isNotEmpty(serviceId)) {
             indexAndValueMap.put(IoTDBIndexes.SERVICE_ID_IDX, serviceId);
         }
-        query = client.addQueryIndexValue(ProfileTaskRecord.INDEX_NAME, query, indexAndValueMap);
+        IoTDBUtils.addQueryIndexValue(ProfileTaskRecord.INDEX_NAME, query, indexAndValueMap);
 
         StringBuilder where = new StringBuilder(" where ");
         if (StringUtil.isNotEmpty(endpointName)) {
-            where.append(ProfileTaskRecord.ENDPOINT_NAME).append(" = \"").append(endpointName).append("\"").append(" and ");
+            where.append(ProfileTaskRecord.ENDPOINT_NAME).append(" = \"")
+                 .append(endpointName).append("\"")
+                 .append(" and ");
         }
         if (Objects.nonNull(startTimeBucket)) {
-            where.append(IoTDBClient.TIME).append(" >= ").append(TimeBucket.getTimestamp(startTimeBucket)).append(" and ");
+            where.append(IoTDBClient.TIME).append(" >= ")
+                 .append(TimeBucket.getTimestamp(startTimeBucket))
+                 .append(" and ");
         }
         if (Objects.nonNull(endTimeBucket)) {
-            where.append(IoTDBClient.TIME).append(" <= ").append(TimeBucket.getTimestamp(endTimeBucket)).append(" and ");
+            where.append(IoTDBClient.TIME).append(" <= ")
+                 .append(TimeBucket.getTimestamp(endTimeBucket))
+                 .append(" and ");
         }
         if (where.length() > 7) {
             int length = where.length();
@@ -73,9 +80,11 @@ public class IoTDBProfileTaskQueryDAO implements IProfileTaskQueryDAO {
         }
         query.append(IoTDBClient.ALIGN_BY_DEVICE);
 
-        List<? super StorageData> storageDataList = client.filterQuery(ProfileTaskRecord.INDEX_NAME, query.toString(), storageBuilder);
+        List<? super StorageData> storageDataList = client.filterQuery(ProfileTaskRecord.INDEX_NAME,
+                                                                       query.toString(), storageBuilder);
         List<ProfileTask> profileTaskList = new ArrayList<>(storageDataList.size());
-        storageDataList.forEach(storageData -> profileTaskList.add(record2ProfileTask((ProfileTaskRecord) storageData)));
+        storageDataList.forEach(storageData ->
+                                        profileTaskList.add(record2ProfileTask((ProfileTaskRecord) storageData)));
         return profileTaskList;
     }
 
@@ -86,27 +95,28 @@ public class IoTDBProfileTaskQueryDAO implements IProfileTaskQueryDAO {
         }
         StringBuilder query = new StringBuilder();
         query.append("select * from ");
-        query = client.addModelPath(query, ProfileTaskRecord.INDEX_NAME);
+        IoTDBUtils.addModelPath(client.getStorageGroup(), query, ProfileTaskRecord.INDEX_NAME);
         Map<String, String> indexAndValueMap = new HashMap<>();
         indexAndValueMap.put(IoTDBIndexes.ID_IDX, id);
-        query = client.addQueryIndexValue(ProfileTaskRecord.INDEX_NAME, query, indexAndValueMap);
+        IoTDBUtils.addQueryIndexValue(ProfileTaskRecord.INDEX_NAME, query, indexAndValueMap);
         query.append(" limit 1").append(IoTDBClient.ALIGN_BY_DEVICE);
 
-        List<? super StorageData> storageDataList = client.filterQuery(ProfileTaskRecord.INDEX_NAME, query.toString(), storageBuilder);
+        List<? super StorageData> storageDataList = client.filterQuery(ProfileTaskRecord.INDEX_NAME,
+                                                                       query.toString(), storageBuilder);
         return record2ProfileTask((ProfileTaskRecord) storageDataList.get(0));
     }
 
     private static ProfileTask record2ProfileTask(ProfileTaskRecord record) {
         return ProfileTask.builder()
-                .id(record.id())
-                .serviceId(record.getServiceId())
-                .endpointName(record.getEndpointName())
-                .startTime(record.getStartTime())
-                .createTime(record.getCreateTime())
-                .duration(record.getDuration())
-                .minDurationThreshold(record.getMinDurationThreshold())
-                .dumpPeriod(record.getDumpPeriod())
-                .maxSamplingCount(record.getMaxSamplingCount())
-                .build();
+                          .id(record.id())
+                          .serviceId(record.getServiceId())
+                          .endpointName(record.getEndpointName())
+                          .startTime(record.getStartTime())
+                          .createTime(record.getCreateTime())
+                          .duration(record.getDuration())
+                          .minDurationThreshold(record.getMinDurationThreshold())
+                          .dumpPeriod(record.getDumpPeriod())
+                          .maxSamplingCount(record.getMaxSamplingCount())
+                          .build();
     }
 }