+++ oap-server/server-storage-plugin/storage-iotdb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/iotdb/query/IoTDBTopNRecordsQueryDAO.java
@@ -18,6 +18,7 @@
 
 package org.apache.skywalking.oap.server.storage.plugin.iotdb.query;
 
+import com.google.common.base.Splitter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -45,6 +46,7 @@ import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.oap.server.storage.plugin.iotdb.IoTDBClient;
 import org.apache.skywalking.oap.server.storage.plugin.iotdb.IoTDBIndexes;
 import org.apache.skywalking.oap.server.storage.plugin.iotdb.IoTDBTableMetaInfo;
+import org.apache.skywalking.oap.server.storage.plugin.iotdb.utils.IoTDBUtils;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -52,22 +54,27 @@ public class IoTDBTopNRecordsQueryDAO implements ITopNRecordsQueryDAO {
     private final IoTDBClient client;
 
     @Override
-    public List<SelectedRecord> readSampledRecords(TopNCondition condition, String valueColumnName, Duration duration) throws IOException {
+    public List<SelectedRecord> readSampledRecords(TopNCondition condition,
+                                                   String valueColumnName,
+                                                   Duration duration)
+            throws IOException {
         StringBuilder query = new StringBuilder();
         query.append("select ").append(TopN.STATEMENT).append(", ").append(valueColumnName)
-                .append(" from ");
-        query = client.addModelPath(query, condition.getName());
+             .append(" from ");
+        IoTDBUtils.addModelPath(client.getStorageGroup(), query, condition.getName());
         Map<String, String> indexAndValueMap = new HashMap<>();
         if (StringUtil.isNotEmpty(condition.getParentService())) {
             final String serviceId = IDManager.ServiceID.buildId(condition.getParentService(), condition.isNormal());
             indexAndValueMap.put(IoTDBIndexes.SERVICE_ID_IDX, serviceId);
         }
-        query = client.addQueryIndexValue(condition.getName(), query, indexAndValueMap);
+        IoTDBUtils.addQueryIndexValue(condition.getName(), query, indexAndValueMap);
 
         StringBuilder where = new StringBuilder(" where ");
         if (Objects.nonNull(duration)) {
-            where.append(IoTDBClient.TIME).append(" >= ").append(TimeBucket.getTimestamp(duration.getStartTimeBucketInSec())).append(" and ");
-            where.append(IoTDBClient.TIME).append(" <= ").append(TimeBucket.getTimestamp(duration.getEndTimeBucketInSec()));
+            where.append(IoTDBClient.TIME).append(" >= ")
+                 .append(TimeBucket.getTimestamp(duration.getStartTimeBucketInSec())).append(" and ");
+            where.append(IoTDBClient.TIME).append(" <= ")
+                 .append(TimeBucket.getTimestamp(duration.getEndTimeBucketInSec()));
         }
         if (where.length() > 7) {
             query.append(where);
@@ -92,8 +99,10 @@ public class IoTDBTopNRecordsQueryDAO implements ITopNRecordsQueryDAO {
                 List<Field> fields = rowRecord.getFields();
                 record.setName(fields.get(1).getStringValue());
 
-                String traceId = fields.get(0).getStringValue().split("\\" + IoTDBClient.DOT + "\"")[traceIdIdx + 1];
-                traceId = client.layerName2IndexValue(traceId);
+                String traceId = Splitter.on(IoTDBClient.DOT + "\"")
+                                         .splitToList(fields.get(0).getStringValue())
+                                         .get(traceIdIdx + 1);
+                traceId = IoTDBUtils.layerName2IndexValue(traceId);
                 record.setRefId(traceId);
 
                 record.setId(record.getRefId());
@@ -111,7 +120,8 @@ public class IoTDBTopNRecordsQueryDAO implements ITopNRecordsQueryDAO {
         // resort by self, because of the select query result order by time.
         if (Order.DES.equals(condition.getOrder())) {
             records.sort((SelectedRecord s1, SelectedRecord s2) ->
-                    Long.compare(Long.parseLong(s2.getValue()), Long.parseLong(s1.getValue())));
+                                 Long.compare(Long.parseLong(s2.getValue()),
+                                              Long.parseLong(s1.getValue())));
         } else {
             records.sort(Comparator.comparingLong((SelectedRecord s) -> Long.parseLong(s.getValue())));
         }